requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
bing-image-downloader==1.1.2
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
flask==3.0.0
flask-cors==4.0.0

